requirements.txt
# Core Requirements
mkdocs-material==9.5.8
mkdocs-git-revision-date-localized-plugin==1.2.1
mkdocs-video==1.5.0
mkdocs-redirects==1.2.1
mkdocs-minify-plugin==0.7.1
mkdocs-static-i18n==1.2.0
git+https://github.com/kamilkrzyskow/gothic-lexer.git@b5ee7868a7e7397cd129ff1f847daa7d39c0e416#egg=gothic_lexer
mkdocs-awesome-pages-plugin==2.9.2

